Code/requirements.txt
numpy
keras
tensorflow
h5py
glob2
gunicorn
Flask
Flask-Uploads
Flask-WTF
Jinja2
Markdown
MarkupSafe
Werkzeug
itsdangerous
db-sqlite3
matplotlib
Pillow
protobuf
opencv-python
opencv-contrib-python
scipy
seaborn
scikit-learn
efficientnet
efficientnet-pytorch
keras-efficientnets


